python/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
mcp>=1.9.0
httpx>=0.28.0
PyJWT>=2.10.0
python-dotenv>=1.0.0
